apps/ai-gateway/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "k12-lms-ai-gateway"
version = "0.1.0"
description = "AI Gateway for the K-12 Planning + LMS platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1.0",
  "httpx>=0.27,<1.0",
  "pydantic-settings>=2.5,<3.0",
  "sentry-sdk[fastapi]>=2.0,<3.0",
  "uvicorn[standard]>=0.30,<1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<10.0",
  "pytest-asyncio>=0.24,<1.0",
  "pytest-cov>=6.0,<7.0",
  "ruff>=0.9,<1.0",
  "mypy>=1.14,<2.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-q --cov=app --cov-report=term-missing --cov-fail-under=80"
